Refactor: For the glory
