+++ ansible_collections/arista/avd/plugins/action/eos_designs_facts.py
@@ -16,10 +16,10 @@ from ansible.plugins.action import ActionBase
 
 from ansible_collections.arista.avd.plugins.plugin_utils.utils import (
     ANSIBLE_ABOVE_2_19,
+    LoaderWrapper,
     get_eos_designs_facts_path,
     get_role_tmp_paths,
     get_templar,
-    load_vaulted_file,
     raise_action_fail,
 )
 
@@ -137,7 +137,8 @@ class ActionModule(ActionBase):
                 raise AnsibleActionFail(message=msg)
 
             # Read, unvault, and parse the JSON file
-            host_hostvars = json.loads(load_vaulted_file(self._loader, file_path))
+            loader = LoaderWrapper(self._loader)
+            host_hostvars = loader.load_file_as_json(file_path)
 
             # Load host hostvars into the AVDDesign data class.
             all_inputs[host] = AVDDesign._from_dict(host_hostvars)
